% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Get UDS/HUF rate}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times failed in the past}
}
\value{
number
}
\description{
Get UDS/HUF rate
}
