% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{unhuf}
\alias{unhuf}
\title{Get HUF value from the forint() return string}
\usage{
unhuf(x)
}
\arguments{
\item{x}{string}
}
\value{
number
}
\description{
Get HUF value from the forint() return string
}
\examples{
unhuf(-3,423.42 Ft)
-3423.42
}
